backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
joblib==1.3.2
scikit-learn==1.4.0
numpy==1.26.3
python-multipart==0.0.6 